Coor_conventions corrected, waypoint nav without any static error

+++ Code/Library/control/estimator.h
@@ -16,7 +16,7 @@ bool filter_init_delta_t;
 uint32_t prev_time;
 
 typedef struct {
-	float state[3][3]; // [z z_speed z_biais] in NED
+	double state[3][3]; // [z z_speed z_biais] in NED
 	float delta_t_filter;
 } Estimator_Data_t;
 
@@ -26,8 +26,8 @@ void e_kalman_init (int axis,float init_p);
 void e_predict (UQuat_t *qe, float a[], float dt);
 void e_kalman_predict (int axis,float accel_meas, float dt);
 void e_kalman_predict_hf(int axis,float accel_meas, float dt);
-void e_kalman_update_position (int axis,float position_meas);
-void e_kalman_update_position_hf(int axis,float position_meas);
+void e_kalman_update_position (int axis, double position_meas);
+void e_kalman_update_position_hf(int axis, double position_meas);
 void e_kalman_update_speed(int axis,float speed_meas);
 void e_kalman_update_speed_hf(int axis,float speed_meas);
